requirements.txt
mediapipe>=0.10.0
opencv-python>=4.5.0
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=1.0.0
joblib>=1.1.0
Pillow>=8.0.0
matplotlib>=3.5.0
